Wheel of POP
Added a wheel of pop page that uses the existing wheel of lunch
javascript class.

Changes to the wheel include:
-added weighting (This also required updating the existing wheel of
lunch page to provide a weighting of 1 for every item)
-added the ability to specify that texts should be perpendicular to the
outer edge of the wheel which works better for short texts
-added the ability to specify that the progress bar should not be
included
-a couple of tweaks to make spins longer for more suspense and drama!

The new wheel of pop's list of items is current hardcoded.